DSFAAP-452: extract CSRF into reusable test helper
