+++ src/server/on-off-farm/controller.test.js
@@ -1,40 +1,10 @@
 import { createServer } from '~/src/server/index.js'
 import { statusCodes } from '~/src/server/common/constants/status-codes.js'
-
-/**
- * @param {OutgoingHttpHeaders} headers
- * @param {string} name
- * @returns {string}
- */
-const findCookie = (headers, name) => {
-  const setCookie = headers['set-cookie']
-
-  /** @type {string[] | undefined} */
-  const cookies = typeof setCookie === 'string' ? [setCookie] : setCookie
-  return cookies?.find((h) => h?.includes(`${name}=`)) ?? ''
-}
-
-/**
- * @param {string} cookieString
- */
-const extractCookieValue = (cookieString) =>
-  cookieString.split('=')[1].split(';')[0]
-
-/**
- * @param {string} payload
- * @param {OutgoingHttpHeaders} headers
- */
-const expectCsrf = (payload, headers) => {
-  const token = extractCookieValue(findCookie(headers, 'crumb'))
-  expect(token).not.toHaveLength(0)
-  expect(payload).toContain(
-    `<input type="hidden" name="crumb" value="${token}"`
-  )
-}
-
-/**
- * @import { OutgoingHttpHeaders } from 'http'
- */
+import {
+  testCrsfProtectedGet,
+  testCrsfProtectedPost,
+  withCrsfProtection
+} from '~/src/server/common/test-helpers/crsf.js'
 
 describe('#onOffFarmController', () => {
   /** @type {Server} */
@@ -50,7 +20,7 @@ describe('#onOffFarmController', () => {
   })
 
   test('Should provide expected response', async () => {
-    const { payload, headers, statusCode } = await server.inject({
+    const { payload, statusCode } = await server.inject({
       method: 'GET',
       url: '/to-or-from-own-premises'
     })
@@ -58,22 +28,19 @@ describe('#onOffFarmController', () => {
     expect(payload).toEqual(
       expect.stringContaining('Are you moving the cattle on or off your farm?')
     )
-    expectCsrf(payload, headers)
     expect(statusCode).toBe(statusCodes.ok)
   })
 
   test('Should process the result and provide expected response', async () => {
-    const { result, statusCode } = await server.inject({
-      method: 'POST',
-      url: '/to-or-from-own-premises',
-      payload: {
-        onOffFarm: 'on',
-        crumb: 'crsf-value'
-      },
-      headers: {
-        Cookie: 'crumb=crsf-value'
-      }
-    })
+    const { result, statusCode } = await server.inject(
+      withCrsfProtection({
+        method: 'POST',
+        url: '/to-or-from-own-premises',
+        payload: {
+          onOffFarm: 'on'
+        }
+      })
+    )
 
     expect(result).toEqual(
       expect.stringContaining('Are you moving the cattle on or off your farm?')
@@ -97,16 +64,12 @@ describe('#onOffFarmController', () => {
   })
 
   test('Should display an error to the user if no value selected', async () => {
-    const { result, statusCode } = await server.inject({
-      method: 'POST',
-      url: '/to-or-from-own-premises',
-      payload: {
-        crumb: 'crsf-value'
-      },
-      headers: {
-        Cookie: 'crumb=crsf-value'
-      }
-    })
+    const { result, statusCode } = await server.inject(
+      withCrsfProtection({
+        method: 'POST',
+        url: '/to-or-from-own-premises'
+      })
+    )
 
     expect(result).toEqual(
       expect.stringContaining(
@@ -119,19 +82,17 @@ describe('#onOffFarmController', () => {
     expect(statusCode).toBe(statusCodes.ok)
   })
 
-  test('should reject if experiencing an apparent CRSF attack (they have their session cookie, but no hidden form value)', async () => {
-    const { statusCode } = await server.inject({
-      method: 'POST',
-      url: '/to-or-from-own-premises',
-      payload: {
-        onOffFarm: 'on'
-      },
-      headers: {
-        Cookie: 'crumb=crsf-value'
-      }
-    })
+  testCrsfProtectedGet(() => server, {
+    method: 'GET',
+    url: '/to-or-from-own-premises'
+  })
 
-    expect(statusCode).toEqual(statusCodes.forbidden)
+  testCrsfProtectedPost(() => server, {
+    method: 'POST',
+    url: '/to-or-from-own-premises',
+    payload: {
+      onOffFarm: 'on'
+    }
   })
 })
